Consensus message format change - Remove SeqNum and timestamp from QC - Change timestamp to ns granularity - Set GENESIS_BLOCK_ID to null

+++ monad-debugger/src/graphql.rs
@@ -235,7 +235,8 @@ impl<'s> From<&'s MonadEventType> for GraphQLMonadEvent<'s> {
                 Self::ControlPanelEvent(GraphQLControlPanelEvent(event))
             }
             MonadEvent::TimestampUpdateEvent(event) => {
-                Self::TimestampEvent(GraphQLTimestampEvent(*event))
+                Self::TimestampEvent(GraphQLTimestampEvent(*event as u64)) // TODO: this is wrong but protobuf is not used in
+                                                                           // protocol and will be deleted
             }
             MonadEvent::StateSyncEvent(event) => Self::StateSyncEvent(GraphQLStateSyncEvent(event)),
         }
